QuantumControlTestUtils

The QuantumControlTestUtils package collects methods that are used for testing and benchmarking within the JuliaQuantumControl organization

Exported Functions

QuantumControlTestUtils.RandomObjects.random_matrix

QuantumControlTestUtils.RandomObjects.random_state_vector

QuantumControlTestUtils.generate_coverage_html

QuantumControlTestUtils.show_coverage

QuantumControlTestUtils.test

Reference

QuantumControlTestUtils.generate_coverage_html — Function

Generate an HTML report for existing coverage data.

generate_coverage_html(path="./"; covdir="coverage", genhtml="genhtml")

creates a folder covdir and use the external genhtml program to write an HTML coverage report into that folder.

QuantumControlTestUtils.show_coverage — Function

Print out a coverage summary from existing coverage data.

show_coverage(path="./src"; sort_by=nothing)

prints a a table showing the tracked files in path, the total number of tracked lines in that file ("Total"), the number of lines with coverage ("Hit"), the number of lines without coverage ("Missed") and the "Coverage" as a percentage.

The coverage data is collected from .cov files in path.

Optionally, the table can be sorted by passing the name of a column to sort_by, e..g. sort_py=:Missed.

QuantumControlTestUtils.test — Function

Run a package test-suite in a subprocess.

test(
    file="test/runtests.jl";
    root=pwd(),
    project="test",
    code_coverage="user",
    show_coverage=(code_coverage == "user"),
    color=<inherit>,
    compiled_modules=<inherit>,
    startup_file=<inherit>,
    depwarn=<inherit>,
    inline=<inherit>,
    check_bounds="yes",
    track_allocation=<inherit>,
    threads=<inherit>,
    genhtml=false,
    covdir="coverage"
)

runs the test suite of the package located at root by running include(file) inside a new julia process.

This is similar to what Pkg.test() does, but differs in the "sandboxing" approach. While Pkg.test() creates a new temporary sandboxed environment, test() uses an existing environment in project (the test subfolder by default). This allows testing against the dev-versions of other packages. It requires that the test folder contains both a Project.toml and a Manifest.toml file.

The test() function also differs from directly including test/runtests.jl in the REPL in that it can generate coverage data and reports (this is only possible when running tests in a subprocess).

If show_coverage is passed as true (default), a coverage summary is shown. Further, if genhtml is true, a full HTML coverage report will be generated in covdir (relative to root). This requires the genhtml executable (part of the lcov package). Instead of true, it is also possible to pass the path to the genhtml exectuable.

All other keyword arguments correspond to the respective command line flag for the julia executable that is run as the subprocess.

This function is intended to be exposed in a project's development-REPL.

QuantumControlTestUtils.RandomObjects.random_matrix — Method

Construct a random matrix.

Ĥ = random_matrix(N; kwargs...)

by default initializes Ĥ as a general complex $N×N$ matrix with a spectral radius of approximately 1.0. Keyword arguments allow to initialize real or complex, Hermitian or non-Hermitian, dense or sparse matrices with arbitrary spectral radius. The non-zero entries in Ĥ will be uniformly distributed around zero, with a range of values that depends on N and the desired spectral radius.

Keyword arguments

density=1.0: A number > 0.0 and ≤ 1.0. Any value < 1.0 implies a sparse matrix where density is the approximate fraction of non-zero elements to total elements

complex=true: Whether the matrix should be complex-valued (default) or real-valued

hermitian=false: Whether the matrix should be general (default) or Hermitian (real eigenvalues)

spectral_radius=1.0: The approximate spectral radius, i.e. maximum absolute eigenvalue. This is according to Girko-Ginibri's circular law, in the limit of large $N$

rng=Random.GLOBAL_RNG: The random number generator to use. The call Random.rand(rng, N, N) must produces a real-valued $N×N$ matrix with elements uniformly distributed between 0 and 1

QuantumControlTestUtils.RandomObjects.random_state_vector — Method

Return a random, normalized Hilbert space state vector of dimension N.

random_state_vector(N; rng=GLOBAL_RNG)