QuantumControlTestUtils

Documentation for QuantumControlTestUtils.

QuantumControlTestUtils.DummyOptimizationResult

QuantumControlTestUtils.QuantumTestLogger

QuantumControlTestUtils.dummy_control_problem

QuantumControlTestUtils.generate_coverage_html

QuantumControlTestUtils.optimize_with_dummy_method

QuantumControlTestUtils.show_coverage

QuantumControlTestUtils.test

QuantumControlTestUtils.DummyOptimizationResult — Type

Result returned by optimize_with_dummy_method.

QuantumControlTestUtils.QuantumTestLogger — Type

Logger that stores all messages in a messages field (array of strings).

The in operator can be used to check whether a string or regex is in any of the recorded messages.

test_logger = QuantumTestLogger()
with_logger(test_logger) do
    #...
end
@test "message" in logger

This is a simplified version of Logging.TestLogger.

QuantumControlTestUtils.dummy_control_problem — Method

Set up a dummy control problem.

problem = dummy_control_problem(;
    N=10, n_objectives=1, n_controls=1, n_steps=50, dt=1.0, sparsity=0.5,
    complex_operators=true, hermitian=true, kwargs...)

Sets up a control problem with random (sparse) Hermitian matrices.

Arguments

N: The dimension of the Hilbert space

n_objectives: The number of objectives in the optimization. All objectives will have the same Hamiltonian, but random initial and target states.

n_controls: The number of controls, that is, the number of control terms in the control Hamiltonian. Each control is an array of random values, normalized on the intervals of the time grid.

n_steps: The number of time steps (intervals of the time grid)

dt: The time step

sparsity: The sparsity of the Hamiltonians, as a number between 0.0 and 1.0. For sparsity=1.0, the Hamiltonians will be dense matrices.

complex_operators: Whether or not the drift/control operators will be complex-valued or real-valued.

hermitian: Whether or not all drift/control operators will be Hermitian matrices.

kwargs: All other keyword arguments are passed on to ControlProblem

QuantumControlTestUtils.generate_coverage_html — Function

Generate an HTML report for existing coverage data.

generate_coverage_html(path="./"; covdir="coverage", genhtml="genhtml")

creates a folder covdir and use the external genhtml program to write an HTML coverage report into that folder.

QuantumControlTestUtils.optimize_with_dummy_method — Method

Run a dummy optimization.

result = optimize(problem, method=:dummymethod)

runs through and "optimization" of the given problem where in each iteration, the amplitude of the guess pulses is diminished by 10%. The (summed) vector norm of the the control serves as the value of the optimization functional.

QuantumControlTestUtils.show_coverage — Function

Print out a coverage summary from existing coverage data.

show_coverage(path="./src"; sort_by=nothing)

prints a a table showing the tracked files in path, the total number of tracked lines in that file ("Total"), the number of lines with coverage ("Hit"), the number of lines without coverage ("Missed") and the "Coverage" as a percentage.

The coverage data is collected from .cov files in path.

Optionally, the table can be sorted by passing the name of a column to sort_by, e..g. sort_py=:Missed.

QuantumControlTestUtils.test — Function

Run a package test-suite in a subprocess.

test(
    file="test/runtests.jl";
    root=pwd(),
    project="test",
    code_coverage="user",
    show_coverage=(code_coverage == "user"),
    color=<inherit>,
    compiled_modules=<inherit>,
    startup_file=<inherit>,
    depwarn=<inherit>,
    inline=<inherit>,
    check_bounds="yes",
    track_allocation=<inherit>,
    threads=<inherit>,
    genhtml=false,
    covdir="coverage"
)

runs the test suite of the package located at root by running include(file) inside a new julia process.

This is similar to what Pkg.test() does, but differs in the "sandboxing" approach. While Pkg.test() creates a new temporary sandboxed environment, test() uses an existing environment in project (the test subfolder by default). This allows testing against the dev-versions of other packages. It requires that the test folder contains both a Project.toml and a Manifest.toml file.

The test() function also differs from directly including test/runtests.jl in the REPL in that it can generate coverage data and reports (this is only possible when running tests in a subprocess).

If show_coverage is passed as true (default), a coverage summary is shown. Further, if genhtml is true, a full HTML coverage report will be generated in covdir (relative to root). This requires the genhtml executable (part of the lcov package). Instead of true, it is also possible to pass the path to the genhtml exectuable.

All other keyword arguments correspond to the respective command line flag for the julia executable that is run as the subprocess.

This function is intended to be exposed in a project's development-REPL.